pointcloudviewer/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
pytest
transformers==4.44.2
einops
timm
opencv-python
tqdm
# flash_attn # If NVIDIA GPU with CUDA is available
# sam2 setup is specific, usually installed via git
# git+https://github.com/facebookresearch/segment-anything-2.git
# git clone https://github.com/facebookresearch/segment-anything-2.git
# cd backend/segment-anything-2
# pip install -e .